refactor(perf): split metrics summary module
